requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
jinja2==3.1.6
python-dotenv==1.1.1
httpx==0.25.2
feedparser==6.0.11
APScheduler==3.11.0
python-telegram-bot==20.7
python-multipart==0.0.6